Create README.md
#Abstract:

In today’s world, computer technology has grown a lot, and to cope up with it we need a more natural way of interacting with computers. The invention of the keyboard and mouse has helped a lot but there are still some situations where we need a more natural way of interaction. Advancement in Human Interaction technology brings a very positive impact in our society as it becomes easy for individuals to interact with computers. Human beings are using a wider range of hand gestures to communicate with each other as it is a mode of non-verbal yet powerful interaction medium. Also, hand gestures provide the most originative and intuitive and more usual way to interact with computers as compared to other body parts. In this paper we made a CNN (Convolutional Neural Network) based model to recognise the static hand gesture images and, with that we have extensively compared our model with several Machine Learning models including SVM,Decision Tree , Naive Bayes, random forest.

#The Rest is mentioned in the Project Report in the Repository.